Fixed failing resource form tests: removed superfluous fields from the form.
The resource form's fields were generated from the Resource model's
fields.  This is unecessary, and now only the required fields are
hard-coded.  This was achieved by copy/pasting from the ckan form
module, and editing as required.  The reason for copy/pasting is that
otherwise, we'd have to wait for the next ckan release.  Copy/pasting
is OK because we'll be moving over to the new forms for DGU soon
anyway, so the duplication is short-lived.