Fixed the trackevent for GA.
When requesting GA to track an event that is a download the event will
not be tracked (in some browsers) where the target of the link is the
same window.  This happens because although the call to _gaq.push will
put the event on the queue, the browser will not have time to process
the queue.  To resolve this, and to avoid opening up tabs and windows
we now instead pause momentarily (200ms) to allow GA to complete the
tracking and then redirect the user to the url they requested.