Fixes #199 sort by title by cleaning the title_string as well as title.
CKAN core lstrips the title in the pkg_dict, but it is the title_string that
is used for search, so this change in search_indexing attempts to clean
up the title string before insertion by lstripping whitespace and removing
puncutation where the title_string leads with it .. i.e. 'Ageing in the UK' dataset.