[393] Removes the GenshiStreamFilter used for extent map - WARNING, read commit msg
Removes the DguDatasetExtentMap as a genshi stream filter and replaces it with two
simpler helpers.  This means that the extent filter is removed from the plugins
in config, and no longer in the setup.py so you will need to re-run the python
setup.py develop before restarting apache.

The config option 'ckan.spatial.dataset_extent_map.routes' is no longer needed.