[229] Adds a last_major_modification extra to the package
When a resource is added, or deleted the last_major_modification is updated to today's date.
If the URL of an existing resource is modified, the same change is made.

Unfortunately CKAN uses the same function name in two different plugins which means that
they can't both be in a single plugin :( meaning we've added two new plugins which will
require a re-setup.py

Display the last_major_modified instead of metadata modified

Added a one-off migration and modified templates