[229] A non-pretty fix to show new extra on the search results page.
When editing a dataset the normal workflow happens and ends up with the package
represented as a dict, and then the resource plugins that we use to are given
instances of the sqlalchemy models making it impossible to add extra within these
plugins. Rather we can add the extras, but we can't add them to the package dict
that is about to be indexed.

This is probably only an issue if we want to add fields to the package/package extras
within the resource plugins.