[541] Fix for javascript re-positioning
Now correctly copies all of the hidden fields into the relevant TD (so that the ID is correctly carried across and not left).  Previously the ID was tied to the position, because of the __pos__ part of the ID and so caused weird behaviour.

Also added an additional_resource__{num}__id as it was missing and I think we're likely to be creating new additional_resources all of the time.