[552] Adds a paster command to remap our resources based on gov.uk remapper
Note: Requires setup.py be re-run

Pulls the repository from github and then after building a local list
of resource urls checks each one in the repo for a match. Uses the ORM.

If there is a match, and the remapper claims it is a move, then we will
update the resource url.  If the remapper claims it is gone then we then
remove the resource, and remove the dataset if there are no more resources.

When run it has a couple of options:
  -n Don't pull the latest code if the repo is already in the /tmp/remapper
     directory.
  -p Just pretend to write to the database, but don't actually do it.
  -c The config file (for the database settings.)

Also contains a minor change to allow the running_stats to order the count
by title instead of count.