[566] Initial attempt but currently has issues.
Need to address how timeseries resources will be managed.
Would be lovely if each timeseries data generated a single resource and the original timeseries resource could be moved to documentation, but I am not sure this is the case.

Needs more investigation into how to handle them and to make sure new resources have the same resource_type as the one that generated it.