[575] Ensure that the url isn't automatically modified when changing title on existing group.
When we have an existing group the following rules apply:

Sysadmin can change title, but url doesn't change. Instead url has a tooltip asking to contact a systemin.

When an admin edits a publisher they can already not edit the title, and the url will be disabled with the previous message about contacting a sysadmin.