[87] Add group abbreviation to search index.
Modifies the search schema and the searchindexing to make sure that group
abbreviations are added to the index, so that searches on the package search
page will return hits for that publisher.  Seems to work with NIO (Northern
Ireland Office) although only returns 4 hits so need to investigate deeper
into why it isn't returning all datasets for that publisher.