[569] Reworks the viewing of large ONS datasets
I've deviated slightly from the plan, and instead have defaulted now to showing just the contents of the latest release (for ONS datasets) and provide a link which will show all of the resources, grouped by the release.  This feels like it is workable as the only real different from a normal dataset is the presence of the View X resources in Y releases link.