[642] Inventory stage 2 changes.
Re-enables inventory items in search, and allows the user to turn them on or off in the results.  They look a little different but definitely need more design work.

Added in some more routes, although this is definitely not required we might re-use the /publisher/ folder structure.

A couple of new plugins to make template handling easier.