[noticket] Moved to OrganizationController and wish I hadn't
A lot of the methods we expect to be in the IGroupForm instance are instead overwritten in the OrganizationController (which is itself a GroupController) and so I've had to move some stuff into the controller.

It may well be that we can remove the IGroupForm entirely as its purpose seems to be replaced with OrganizationController subclasses.  But that is tomorrow's task, now I am going to sit in the corner and cry for a little bit.