[1014] JS Fix for pagination on openspending browse.
Unfortunately the layout of the HTML served from OpenSpending isn't ideal and we are missing some CSS.

As a temporary measure (and as a test) we wait 500ms after loading and try and add the relevant classes ourselves.  Only seems to work if you have a local cache of the relevant openspending js. Sometimes doesn't work first time, but now sure how to wait until the remote thing has loaded.