[1067] Updated theme-categorization for LGA
Added two more scoring methods to the theme categorization to handle
LGA datasets that have been harvested to give them themes based on
the service and function URLs provided in metadata. The IDs are
taken directly from the function/service URLs.  If we find the
appropriate extras the functions are scored very highly, whilst
services are scored as more of a hint.

The list of function/service IDs was generated by ckanext-dgu-local/ckanext/dgulocal/tools/import_themes.py using the data file in the data/ folder.