Allows publishers to be marked as closed.
Adds a new schema field called closed which is used to denote a closed
publisher.  The edit form has been jinjafied and now has a new checkbox
which sets the closed state of the publisher.  To get this to work we
add a new version of the core CKAN boolean_validator that also allows
the Missing value to be treated as false.