Adds a 'replaced_by' field to the publisher form.
Now when marking a publisher as closed, it is possible to select one or
more publishers who have now 'replaced' the current one.  It isn't
required and there's no easy way (that I can see) to limit the number of
replacements, but it seems to work.

The message generated for closed publishers on publisher_read and
publisher_edit is done from a helper.