Adds a new resource/X/api page with info on using the page
Also adds an API button on the resource_read page where there is an API
for the resource.  The API page shows the schema for the API and gives a
couple of examples and explains how to use it.  I may yet switch this to
use swagger so that users can actually try out the API on the page
(http://swagger.io).