Enable the datapusher status page.
It isn't JUST the datapusher status page, it's part of the tabbed
interface in ckan core templates - but it's the only bit we want.

When a sysadmin goes to a resource page they'll see a status button
which'll show them:

a. The status wrt datastore
b. A button to push the resource through the pusher and to the
datastore.  This still won't work for non-NII packages