Tidy up code for deciding the Content-Type of cached resources
Unless you pass the content_type kwarg to the FileApp constructor it ignores
the Content-Type header and makes a guess. The existing code intends to use the
format decided by ckanext-qa, however for zipped resources that extension
returns the format inside the zip, so should not be used for deciding the
Content-Type to return. Therefore as this code is not used in deciding which
Content-Type will be returned to the user and would reqire major changes to
ckanext-qa, I am removing it to avoid confusion.