Fixes the inventory tests.
Requires DB to be postgres for the CTE around groups.  Was using the
wrong form on a lot of pages, the first one is our search box.

One thing that is problematic is that after a POST we redirect, and
normally you'd call follow() on the response to get to the actual
content - but unfortunately the Location header generated appears to be
invalid.