Add links from datasets to collections.
For the given dataset, autolink will find all the publications it has
links to, and if there are enough links to constitute a majority for the
collection - creates a link.

Unfortunately it requires a migration (if you have an already working
system) because we added a unique identifier to the collection model.

    alter table link alter column govuk_id type text;
    alter table collection add column govuk_id text default 0;