Adds a command to repair timeseries datasets.
resources to time-series, but left some as individual-resources. This
now leaves the dataset uneditable because of validation, and no way of
setting them as time-series as they do not show up in the edit form for
the dataset.

This tool will convert all the individual-resources to time-series if
there are any time-series resources already.