Fixes #275 by importing individual resources into timeseries table.
In cases where something has gone terribly wrong, we may end up with a
tiemseries dataset, but with some individual_resources.  This will cause
a validation error.

When this occurs, if we are in timeseries mode, we will include the
individual_resources in the timeseries table so that the validation
error is missing date, rather than a mixed resource type warning.  We
also do not populate the invidual_resource table, it will be sorted out
on the client if we switch back, but adding it will trigger the wrong
validation again.