Do not process harvested items.
For some reason, if a harvested item fails on validating, for example,
spatial data, the revision gets messed up and it is no longer available.

This commit will just skip the harvested items, which will eventually
overwrite any changes, but doesn't solve the potential issue with the
changes via ckanapi.