Fix search relevance
CKAN doesn't process 'rank desc' properly in search results and so we
end up with just an alphabet ordering.  The default is relevance so we
just remove the value for sort.

This means going from Relevance to Popular and back to Relevance gives
you the same results as initially (after searching for a term), rather
than ordering by title.

Fixes #329