Fixes flicker with Google Analytics on page
When google analytics was added to the very last part of the page, the
chosen selects would flicker on recent copies of Chrome.  This appears
to be because the GA code was inserting a script tag (and waiting for
the JS to load asynchronously).  Removing the GA snippet, or moving
above the chosen call solves this problem.  It isn't possible to move
the chosen call to after the existing snippets as it has to be called
after $(document).ready().

This commit resolves the issue by providing a helper to load the analytics
snippet (now renamed to googleanalytics.snippet ) in the head section of
the page,