Remove the number as they are confusing.
Showing number in the facet dropdowns is confusing, it isn't clear how
it will affect the total results (although it is showing the new total
result value if you select that item).  Also, once selected it shows in
the input field until the page finishes loading, which leads to more
confusion when it disappears - after page load.