Surface registers
Adds a tickbox in the usual place to denote that a dataset is a
register, the result is then boosted much higher when the register
extra is set to true.

Removes the confusing NII box, and replaces with a Register box with
which we can determine if that, or the colour hilight is a better
approach to making it more obvious that it is a register.

Requires a schema change and a re-index